requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
coverage==4.5.1
coveralls==1.5.0
docopt==0.6.2
Flask==1.0.2
Flask-Script==2.0.6
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
psycopg2-binary==2.7.5
PyJWT==1.6.4
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
